Move some utility methods and types into features/identities
