+++ src/features/identities/index.ts
@@ -1,1 +1,3 @@
 export * from './slice'
+export * from './types'
+export * from './utils'
