+++ src/features/identities/slice/index.ts
@@ -1,10 +1,10 @@
 import type { PayloadAction } from '@reduxjs/toolkit'
 import { createSlice } from '@reduxjs/toolkit'
 import { logout } from 'features/auth'
+import { Account } from 'features/identities'
 import { Sentry } from 'features/telemetry'
 import { cloneDeep } from 'lodash'
 
-import { Account } from 'api/types'
 import { RootState } from 'reduxStore/types'
 
 export interface IdentitiesState {
