+++ src/features/profiles/hooks/useWatchPublicProfileChanges.ts
@@ -1,11 +1,10 @@
-import { selectSelectedAccount } from 'features/identities'
+import { Account, selectSelectedAccount } from 'features/identities'
 import { PublicProfile, selectSelectedPublicProfile } from 'features/profiles'
 import { emitter } from 'helpers/emitter'
 import { useEffect } from 'react'
 import { useSelector } from 'react-redux'
 
 import AccountManager from 'api/AccountManager'
-import { Account } from 'api/types'
 
 export function useWatchPublicProfileChanges() {
   const publicProfile: PublicProfile = useSelector(selectSelectedPublicProfile)
